[AN] 컨벤션 변경에 따른 리소스 파일 이름명 수정 및 클래스명 통일 (#101)
* refactor: 바인딩 어댑터 패키지 위치 변경 및 북마크 아이콘 리소스명 변경

* refactor: `Place` 클래스 이름 변경 및 파일 이동

- `Place` 클래스의 이름을 `PlaceUiModel`로 변경
- `PlaceUiModel` 및 `PlaceCategory`를 `uimodel` 패키지에서 `model` 패키지로 이동
- `PlaceUiModel`에 `isBookmarked` 프로퍼티 추가

* feat: MainActivity 바텀 네비게이션 기능 구현 및 Fragment 연결

- NewsFragment 인스턴스 생성
- 바텀 네비게이션 아이템 클릭 시 해당 Fragment로 전환하는 로직 추가
  - Home, Schedule, Map, News 메뉴에 각각 Fragment 연결
- Fragment 태그 상수 추가 (TAG_PLACE_LIST_FRAGMENT, TAG_NEW_FRAGMENT)
- 초기 Fragment를 HomeFragment로 설정하도록 수정
- 사용하지 않는 .gitkeep 파일 삭제

* refactor: BindingAdapter 파일 통합

- PlaceListBindingAdapter.kt 파일을 삭제하고 해당 내용을 presentation/common/BindingAdapter.kt 로 이동

* refactor: NewsFragment, NoticeUiModel 경로 변경 및 관련 코드 수정

- NewsFragment: presentation.news.notice -> presentation.news
- NoticeUiModel: presentation.news.notice -> presentation.news.notice.model
- PlaceListViewHolder, PlaceListAdapter: presentation.placeList -> presentation.placeList.adapter
- 레이아웃 파일의 tools:context 및 data binding 경로 수정
- MainActivity의 NewsFragment import 경로 수정
- 레이아웃 파일 내 뷰 속성 순서 변경

* refactor: ScheduleAdapter DiffUtil 상수명 변경

- ScheduleAdapter의 DiffUtil 상수명을 `DIFF_UTIL_CALLBACK`에서 `DIFF_UTIL`로 변경

* refactor: ScheduleEventMapper 제거 및 확장 함수 위치 변경

* feat: drawable 리소스 파일명 변경
