+++ android/app/src/main/java/com/daedan/festabook/presentation/placeList/adapter/PlaceListViewHolder.kt
@@ -1,22 +1,23 @@
-package com.daedan.festabook.presentation.placeList
+package com.daedan.festabook.presentation.placeList.adapter
 
+import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import androidx.viewbinding.ViewBinding
 import com.daedan.festabook.databinding.ItemPlaceListBinding
 import com.daedan.festabook.databinding.ItemPlaceListHeaderBinding
-import com.daedan.festabook.presentation.placeList.uimodel.Place
+import com.daedan.festabook.presentation.placeList.model.PlaceUiModel
 
 sealed class PlaceListViewHolder(
     binding: ViewBinding,
 ) : RecyclerView.ViewHolder(binding.root) {
-    abstract fun bind(place: Place)
+    abstract fun bind(placeUiModel: PlaceUiModel)
 
     class PlaceViewHolder private constructor(
         private val binding: ItemPlaceListBinding,
     ) : PlaceListViewHolder(binding) {
-        override fun bind(place: Place) {
-            binding.place = place
+        override fun bind(placeUiModel: PlaceUiModel) {
+            binding.place = placeUiModel
         }
 
         companion object {
@@ -25,7 +26,7 @@ sealed class PlaceListViewHolder(
             fun of(parent: ViewGroup): PlaceViewHolder =
                 PlaceViewHolder(
                     ItemPlaceListBinding.inflate(
-                        android.view.LayoutInflater.from(parent.context),
+                        LayoutInflater.from(parent.context),
                         parent,
                         false,
                     ),
@@ -36,7 +37,7 @@ sealed class PlaceListViewHolder(
     class Header private constructor(
         binding: ItemPlaceListHeaderBinding,
     ) : PlaceListViewHolder(binding) {
-        override fun bind(place: Place) = Unit
+        override fun bind(placeUiModel: PlaceUiModel) = Unit
 
         companion object {
             val VIEW_TYPE = 1
@@ -44,7 +45,7 @@ sealed class PlaceListViewHolder(
             fun of(parent: ViewGroup): Header =
                 Header(
                     ItemPlaceListHeaderBinding.inflate(
-                        android.view.LayoutInflater.from(parent.context),
+                        LayoutInflater.from(parent.context),
                         parent,
                         false,
                     ),
