+++ android/app/src/main/java/com/daedan/festabook/presentation/placeList/model/PlaceUiModel.kt
@@ -1,10 +1,11 @@
-package com.daedan.festabook.presentation.placeList.uimodel
+package com.daedan.festabook.presentation.placeList.model
 
-data class Place(
+data class PlaceUiModel(
     val id: Long,
     val imageUrl: String,
     val category: PlaceCategory,
     val title: String,
     val description: String,
     val location: String,
+    val isBookmarked: Boolean = false,
 )
