+++ android/app/src/main/java/com/daedan/festabook/presentation/schedule/adapter/ScheduleItemViewHolder.kt
@@ -9,10 +9,9 @@ import androidx.recyclerview.widget.RecyclerView
 import com.daedan.festabook.R
 import com.daedan.festabook.databinding.ItemScheduleTabPageBinding
 import com.daedan.festabook.presentation.schedule.OnBookmarkCheckedListener
-import com.daedan.festabook.presentation.schedule.mapper.toKoreanString
-import com.daedan.festabook.presentation.schedule.mapper.toPx
 import com.daedan.festabook.presentation.schedule.model.ScheduleEventUiModel
 import com.daedan.festabook.presentation.schedule.model.ScheduleEventUiStatus
+import com.daedan.festabook.presentation.schedule.model.toKoreanString
 
 class ScheduleItemViewHolder(
     private val binding: ItemScheduleTabPageBinding,
@@ -42,7 +41,7 @@ class ScheduleItemViewHolder(
                     textColor = gray400,
                     backgroundResId = null,
                 )
-                setupScheduleEventTimeLineCircleIcon(R.drawable.ic_schedule_circle_gray300)
+                setupScheduleEventTimeLineCircleIcon(R.drawable.ic_circle_gray300)
                 setupScheduleEventContentsColor(
                     titleColor = gray400,
                     timeColor = gray400,
@@ -52,7 +51,7 @@ class ScheduleItemViewHolder(
             }
 
             ScheduleEventUiStatus.ONGOING -> {
-                val borderColor = R.drawable.bg_schedule_stroke_blue400_radius_10dp
+                val borderColor = R.drawable.bg_stroke_blue400_radius_10dp
                 binding.clScheduleEventCard.setBackgroundResource(borderColor)
                 setupScheduleEventStatusText(
                     context = context,
@@ -60,7 +59,7 @@ class ScheduleItemViewHolder(
                     textColor = gray050,
                     backgroundResId = R.drawable.bg_gray900_radius_6dp,
                 )
-                setupScheduleEventTimeLineCircleIcon(R.drawable.ic_schedule_circle_blue400)
+                setupScheduleEventTimeLineCircleIcon(R.drawable.ic_circle_blue400)
                 setupScheduleEventContentsColor(
                     titleColor = gray900,
                     timeColor = gray500,
@@ -70,7 +69,7 @@ class ScheduleItemViewHolder(
             }
 
             ScheduleEventUiStatus.UPCOMING -> {
-                val borderColor = R.drawable.bg_schedule_stroke_1dp_radius_10dp
+                val borderColor = R.drawable.bg_stroke_green400_radius_10dp
                 binding.clScheduleEventCard.setBackgroundResource(borderColor)
                 setupScheduleEventStatusText(
                     context = context,
@@ -78,7 +77,7 @@ class ScheduleItemViewHolder(
                     textColor = gray900,
                     backgroundResId = R.drawable.bg_stroke_gray900_radius_6dp,
                 )
-                setupScheduleEventTimeLineCircleIcon(R.drawable.ic_schedule_circle_green400)
+                setupScheduleEventTimeLineCircleIcon(R.drawable.ic_circle_green400)
                 setupScheduleEventContentsColor(
                     titleColor = gray900,
                     timeColor = gray500,
@@ -138,3 +137,5 @@ class ScheduleItemViewHolder(
         }
     }
 }
+
+fun Int.toPx(context: Context): Int = (this * context.resources.displayMetrics.density).toInt()
