+++ android/app/src/main/res/drawable/bg_gray400_radius_50dp.xml
