+++ android/app/src/main/res/drawable/bg_stroke_blue400_radius_10dp.xml
