+++ android/app/src/main/res/drawable/bg_stroke_green400_radius_10dp.xml
