+++ android/app/src/main/res/drawable/ic_circle_blue400.png
