+++ android/app/src/main/res/drawable/ic_circle_gray300.png
