--- android/app/src/main/res/drawable/ic_place_list_location.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="14dp"
-    android:height="14dp"
-    android:viewportWidth="14"
-    android:viewportHeight="14">
-
-    <path
-        android:fillColor="@color/gray500"
-        android:pathData="M7.00008 8.49967C8.28691 8.49967 9.33341 7.45317 9.33341 6.16634C9.33341 4.87951 8.28691 3.83301 7.00008 3.83301C5.71325 3.83301 4.66675 4.87951 4.66675 6.16634C4.66675 7.45317 5.71325 8.49967 7.00008 8.49967ZM7.00008 4.99967C7.6435 4.99967 8.16675 5.52292 8.16675 6.16634C8.16675 6.80976 7.6435 7.33301 7.00008 7.33301C6.35666 7.33301 5.83341 6.80976 5.83341 6.16634C5.83341 5.52292 6.35666 4.99967 7.00008 4.99967Z" />
-
-    <path
-        android:fillColor="@color/gray500"
-        android:pathData="M6.66164 12.7252C6.76043 12.7955 6.87869 12.8333 6.99997 12.8333C7.12125 12.8333 7.23951 12.7955 7.33830 12.7252C7.51564 12.5997 11.6836 9.59033 11.6666 5.83366C11.6666 3.26058 9.57305 1.16699 6.99997 1.16699C4.42689 1.16699 2.33330 3.26058 2.33330 5.83074C2.31639 9.59033 6.48430 12.5997 6.66164 12.7252ZM6.99997 2.33366C8.93022 2.33366 10.5 3.90341 10.5 5.83658C10.5122 8.42541 7.94030 10.75 6.99997 11.5124C6.06022 10.7494 3.48772 8.42424 3.49997 5.83366C3.49997 3.90341 5.06972 2.33366 6.99997 2.33366Z" />
-
-</vector>
