+++ android/app/src/main/res/drawable/selector_bg_gray900_radius_20.xml
@@ -1,4 +1,4 @@
 <selector xmlns:android="http://schemas.android.com/apk/res/android">
-    <item android:drawable="@drawable/schedule_bg_gray900_radius_20dp"  android:state_selected="true" />
+    <item android:drawable="@drawable/bg_gray900_radius_20dp"  android:state_selected="true" />
     <item android:drawable="@android:color/transparent" />
 </selector>
