+++ android/app/src/main/res/layout/item_place_list_header.xml
@@ -12,7 +12,7 @@
             android:layout_width="32dp"
             android:layout_height="4dp"
             android:layout_marginTop="12dp"
-            android:background="@drawable/div_place_list"
+            android:background="@drawable/bg_gray400_radius_50dp"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent" />
