+++ android/app/src/main/res/layout/item_schedule_tab_page.xml
@@ -115,7 +115,7 @@
                 android:layout_marginBottom="20dp"
                 android:contentDescription="@string/all_book_mark"
                 android:onClick="@{() -> onBookmarkCheckedListener.onBookmarkChecked(scheduleEvent.id)}"
-                android:src="@drawable/ic_bookmark"
+                android:src="@drawable/ic_bookmark_normal"
                 app:layout_constraintBottom_toBottomOf="@id/cl_schedule_event_card"
                 app:layout_constraintEnd_toEndOf="@id/cl_schedule_event_card" />
 
@@ -140,7 +140,7 @@
             android:layout_marginEnd="12dp"
             android:contentDescription="@string/schedule_circle"
             android:elevation="11dp"
-            android:src="@drawable/ic_schedule_circle_gray300"
+            android:src="@drawable/ic_circle_gray300"
             app:layout_constraintBottom_toBottomOf="@id/cl_schedule_event_card"
             app:layout_constraintEnd_toStartOf="@id/cl_schedule_event_card"
             app:layout_constraintStart_toStartOf="parent"
